Work on connector migrations
