+++ arroyo-api/src/connection_tables.rs
@@ -11,13 +11,13 @@ use futures_util::stream::Stream;
 use serde_json::{json, Value};
 use std::convert::Infallible;
 use tokio::sync::mpsc::channel;
-use tokio_stream::StreamExt;
 use tokio_stream::wrappers::ReceiverStream;
+use tokio_stream::StreamExt;
 use tracing::warn;
 
 use arroyo_connectors::confluent::ConfluentProfile;
-use arroyo_connectors::kafka::{KafkaConfig, KafkaTable, SchemaRegistry};
 use arroyo_connectors::connector_for_type;
+use arroyo_connectors::kafka::{KafkaConfig, KafkaTable, SchemaRegistry};
 use arroyo_formats::avro;
 use arroyo_formats::json_schema::to_arrow;
 use arroyo_operator::connector::ErasedConnector;
@@ -35,13 +35,13 @@ use arroyo_types::raw_schema;
 
 use crate::rest::AppState;
 use crate::rest_utils::{
-    ApiError, authenticate, bad_request, BearerAuth, client, ErrorResp, log_and_map,
-    not_found, paginate_results, required_field, validate_pagination_params,
+    authenticate, bad_request, client, log_and_map, not_found, paginate_results, required_field,
+    validate_pagination_params, ApiError, BearerAuth, ErrorResp,
 };
 use crate::{
-    AuthData, handle_db_error,
-    handle_delete,
-    queries::api_queries::{self, DbConnectionTable}, to_micros,
+    handle_db_error, handle_delete,
+    queries::api_queries::{self, DbConnectionTable},
+    to_micros, AuthData,
 };
 
 async fn get_and_validate_connector<E: GenericClient>(
@@ -184,7 +184,9 @@ pub(crate) async fn test_connection_table(
 
     let stream = ReceiverStream::new(rx);
 
-    Ok(Sse::new(stream.map(|msg| Ok(Event::default().json_data(msg).unwrap()))))
+    Ok(Sse::new(
+        stream.map(|msg| Ok(Event::default().json_data(msg).unwrap())),
+    ))
 }
 
 fn get_connection_profile(
