+++ arroyo-connectors/src/filesystem.rs
@@ -5,15 +5,15 @@ use std::collections::HashMap;
 use std::convert::Infallible;
 use typify::import_types;
 
+use arroyo_operator::connector::Connection;
 use arroyo_rpc::api_types::connections::{
     ConnectionProfile, ConnectionSchema, ConnectionType, TestSourceMessage,
 };
 use arroyo_rpc::formats::Format;
 use arroyo_rpc::OperatorConfig;
 use serde::{Deserialize, Serialize};
-use arroyo_operator::connector::Connection;
 
-use crate::{EmptyConfig, pull_opt, pull_option_to_i64};
+use crate::{pull_opt, pull_option_to_i64, EmptyConfig};
 
 use arroyo_operator::connector::Connector;
 
@@ -64,9 +64,7 @@ impl Connector for FileSystemConnector {
                 done: true,
                 message: "Successfully validated connection".to_string(),
             };
-            tx.send(message)
-                .await
-                .unwrap();
+            tx.send(message).await.unwrap();
         });
     }
