+++ arroyo-connectors/src/impulse/table.json
