+++ arroyo-connectors/src/polling_http/table.json
