+++ arroyo-connectors/src/redis/operator/mod.rs
@@ -1,0 +1,1 @@
+pub mod sink;
