+++ arroyo-connectors/src/redis/profile.json
