+++ arroyo-connectors/src/redis/table.json
