+++ arroyo-worker/src/network_manager.rs
@@ -22,9 +22,9 @@ use tokio::{
     sync::mpsc::{Receiver, Sender},
 };
 
+use arroyo_operator::context::QueueItem;
 use tokio::time::{interval, Interval};
 use tokio_stream::StreamExt;
-use arroyo_operator::context::QueueItem;
 
 use arroyo_operator::inq_reader::InQReader;
 
@@ -468,7 +468,7 @@ mod test {
     use std::time::SystemTime;
     use std::{pin::Pin, time::Duration};
 
-    use arroyo_types::{ArrowMessage, CheckpointBarrier, SignalMessage, to_nanos};
+    use arroyo_types::{to_nanos, ArrowMessage, CheckpointBarrier, SignalMessage};
     use tokio::{sync::mpsc::channel, time::timeout};
 
     use crate::network_manager::{MessageType, Quad};
